+++ test/clj_reload/parse_test.clj
@@ -105,8 +105,10 @@ Unexpected :require form: [789 a b c]
 
 (deftest scan-impl-test
   (let [{files :files'
-         nses  :namespaces'} (binding [util/*log-fn* nil]
-                               (@#'reload/scan-impl nil ["fixtures"] #".*\.cljc?" 0))]
+         nses  :namespaces'} (binding [reload/*config* {:dirs ["fixtures"]
+                                                        :files #".*\.cljc?"}
+                                       util/*log-fn*  nil]
+                               (@#'reload/scan-impl nil 0))]
     (testing "no-ns"
       (is (= '#{}
             (get-in files [(io/file "fixtures/core_test/no_ns.clj") :namespaces]))))
@@ -151,9 +153,10 @@ Unexpected :require form: [789 a b c]
             (get-in nses ['double :ns-files]))))
 
     (testing "custom-file-types"
-      (let [custom-file-type-re #".*\.(?:cljc?|repl)"
-            {files-custom :files'} (binding [util/*log-fn* nil]
-                                     (@#'reload/scan-impl nil ["fixtures"] custom-file-type-re 0))]
+      (let [{files-custom :files'} (binding [reload/*config* {:dirs  ["fixtures"]
+                                                              :files #".*\.(?:cljc?|repl)"}
+                                             util/*log-fn* nil]
+                                     (@#'reload/scan-impl nil 0))]
         (is (nil?
               (get-in files [(io/file "fixtures/core_test/custom_file_type.repl") :namespaces])))
         (is (= '#{custom-file-type}